Stack fixed
The problem was that when the stack was only 1 event high, the event
wasn't being zero'd out, so eventCount picked it up as a filed value,
effectively duping the first action performed.  I also added code to
zero out the last value that was being moved down in the stack because
otherwise eventCount would pick that one up too, duping the last event
entered.