Fixed the joystick problem
The problem was that far too many events were being generated, much
faster than they were being consumed.  I fixed it by creating
pushEvent_replace, which finds the highest event on the stack with the
same type as the one we want to push and just changes the 'data' member
field.